+++ app/app.tsx
@@ -54,11 +54,12 @@ import { openActiveRide } from "./utils/helpers/ride-helpers"
 import { GestureHandlerRootView } from "react-native-gesture-handler"
 import { createModalStack, ModalProvider } from "react-native-modalfy"
 import { TipThanksModal } from "./screens/settings/components/tip-thanks-modal"
+import { RouteListWarningModal } from "./screens/route-list/components/route-list-warning-modal"
 enableScreens()
 
 export const queryClient = new QueryClient()
 
-const modalConfig = { TipThanksModal }
+const modalConfig = { TipThanksModal, RouteListWarningModal }
 const defaultOptions = { backdropOpacity: 0.6 }
 
 const stack = createModalStack(modalConfig, defaultOptions)
